Moved player related events out of main.rs
Moved player related events out of main.rs (Yeah we just now finally figured this out)
