+++ src/main.rs
@@ -21,13 +21,12 @@ static GLOBAL: mimalloc::MiMalloc = mimalloc::MiMalloc;
 ///////////////////////////////////////////
 
 // Imported some third party crates
-use http::status;
 use serde_json::{json, Value};
 use socketioxide::extract::{AckSender, Bin, Data, SocketRef};
 use std::sync::{Arc, Mutex};
 use tokio::{main, task::spawn};
 use tracing::{debug, info};
-use viz::{future::ok, handler::ServiceHandler, serve, Response, Result, Router, Request, Body};
+use viz::{handler::ServiceHandler, serve, Response, Result, Router, Request, Body};
 
 
 // Import some custom crates from the crates folder in /src
@@ -40,6 +39,7 @@ use PebbleVault;
 // will be very bad but should be fine for now)             //
 //////////////////////////////////////////////////////////////
 use structs::*;
+use players::*;
 
 /////////////////////////////////////
 // Import the modules we will need //
@@ -48,6 +48,7 @@ use structs::*;
 mod events;
 mod macros;
 mod structs;
+mod players;
 mod subsystems;
 
 ///////////////////////////////////////////////////////////////
@@ -57,44 +58,8 @@ mod subsystems;
 ///////////////////////////////////////////////////////////////
 
 fn on_connect(socket: SocketRef, Data(data): Data<Value>, players: Arc<Mutex<Vec<Player>>>) {
-    // Update the parsing functions to handle the data without Object wrappers
-
-    fn parse_rotation(parse: &Value) -> (f64, f64, f64, f64) {
-        (
-            parse_f64(&parse["w"]).unwrap_or(0.0),
-            parse_f64(&parse["x"]).unwrap_or(0.0),
-            parse_f64(&parse["y"]).unwrap_or(0.0),
-            parse_f64(&parse["z"]).unwrap_or(0.0),
-        )
-    }
-
-    fn parse_xyz(parse: &Value) -> (f64, f64, f64) {
-        (
-            parse_f64(&parse["x"]).unwrap_or(0.0),
-            parse_f64(&parse["y"]).unwrap_or(0.0),
-            parse_f64(&parse["z"]).unwrap_or(0.0),
-        )
-    }
-
-    fn parse_xy(parse: &Value) -> (f64, f64) {
-        (
-            parse_f64(&parse["x"]).unwrap_or(0.0),
-            parse_f64(&parse["y"]).unwrap_or(0.0),
-        )
-    }
-
-    fn parse_f64(n: &Value) -> Result<f64, std::io::Error> {
-        n.as_f64().ok_or_else(|| std::io::Error::new(std::io::ErrorKind::InvalidData, "Invalid f64 value"))
-    }
-
-
     let id = socket.id.as_str();
     println!("Starting subsystems for player: {}", id);
-
-    ///////////////////////////////////////////////////////////////////
-    //
-    //
-    ///////////////////////////////////////////////////////////////////
 
     // Authenticate the user
     let player =  Player {
@@ -132,140 +97,31 @@ fn on_connect(socket: SocketRef, Data(data): Data<Value>, players: Arc<Mutex<Vec
     //  file                                               //
     /////////////////////////////////////////////////////////
     
-    //  define_event!(socket, 
-    //      "test", events::test::main(),
-    //      );
 
-    let players_clone_updateloc: Arc<Mutex<Vec<Player>>> = Arc::clone(&players);
-    let players_clone: Arc<Mutex<Vec<Player>>> = Arc::clone(&players);
-    socket.on(
-        "updatePlayerLocation",
-        move |socket: SocketRef, Data::<Value>(data), Bin(bin)| {
-            println!(
-                "Received event: UpdatePlayerLocation with data: {:?} and bin: {:?}",
-                data, bin
-            );
-        
-            // Extract location from data
-            if let Some(transform) = data.get("transform").and_then(|t| t.as_object()) {
-                let mut players: std::sync::MutexGuard<Vec<Player>> = players_clone_updateloc.lock().unwrap();
-                if let Some(player) = players.iter_mut().find(|p: &&mut Player| p.socket.id == socket.id)
-                {
-                    // Do the actual parsing
-                    if let (Some(rotation), Some(translation), Some(scale3d)) = (
-                        transform.get("rotation"),
-                        transform.get("translation"),
-                        transform.get("scale3D")
-                    ) {
-                        let (rot_w, rot_x, rot_y, rot_z) = parse_rotation(rotation);
-                        let (trans_x, trans_y, trans_z) = parse_xyz(translation);
-                        let (scale3d_x, scale3d_y, scale3d_z) = parse_xyz(scale3d);
-                    
-                        // Create or update the transform
-                        let mut transform = player.transform.take().unwrap_or_else(|| structs::Transform {
-                            rotation: None,
-                            translation: None,
-                            scale3D: structs::Scale3D { x: 1.0, y: 1.0, z: 1.0 },
-                            location: None,
-                        });
-                    
-                        // Update rotation
-                        transform.rotation = Some(structs::Rotation { w: rot_w, x: rot_x, y: rot_y, z: rot_z });
-                    
-                        // Update translation
-                        let new_translation = structs::Translation { x: trans_x, y: trans_y, z: trans_z };
-                        transform.translation = Some(new_translation.clone());
-                        transform.location = Some(new_translation);
-                    
-                        // Update scale3D
-                        transform.scale3D = structs::Scale3D { x: scale3d_x, y: scale3d_y, z: scale3d_z };
-                    
-                        // Update the player's transform
-                        player.transform = Some(transform);
-                    
-                        // Parse player movement axis values
-                        if let Some(move_action) = data.get("move Action Value") {
-                            let (mv_action_value_x, mv_action_value_y) = parse_xy(move_action);
-                            player.moveActionValue = Some(structs::MoveActionValue { x: mv_action_value_x, y: mv_action_value_y });
-                        }
-                    
-                        // Print a debug statement
-                        println!("Updated player location: {:?}", player);
-                    } else {
-                        println!("Invalid transform data structure");
-                    }
-                } else {
-                    println!("Player not found: {}", socket.id);
-                }
-            } else {
-                println!("Failed to parse location: transform field not found or is not an object");
-            }
-        
-            // Send a reply containing the correct data
-            socket.bin(bin).emit("messageBack", data).ok();
-        },
+    // Register events using the definevent macro
+    define_event!(socket, 
+        "test", events::test::main(),
     );
 
-    /////////////////////////////////////////////////////////////////////////
-    //  Client sends this message when they need a list of online players  //
-    /////////////////////////////////////////////////////////////////////////
+    // Register events using the socketioxide API directly
+    let players_clone = Arc::clone(&players);
+    socket.on("updatePlayerLocation", move |s: SocketRef, d: Data<Value>| {
+        update_player_location(s, d, players_clone.clone())
+    });
 
-    socket.on(
-        "getOnlinePlayers",
-        move |socket: SocketRef, _: Data<Value>, _: Bin| {
-            info!("Responding with online players list");
-            let players: std::sync::MutexGuard<Vec<Player>> = players_clone.lock().unwrap();
-            let online_players_json = serde_json::to_value(
-                players
-                    .iter()
-                    .map(|player| json!({ "id": player.socket.id }))
-                    .collect::<Vec<_>>(),
-            )
-            .unwrap();
-            debug!("Player Array as JSON: {}", online_players_json);
-            socket.emit("onlinePlayers", online_players_json).ok();
-        },
-    );
+    let players_clone = Arc::clone(&players);
+    socket.on("getOnlinePlayers", move |s: SocketRef| {
+        get_online_players(s, players_clone.clone())
+    });
 
-    let players_clone: Arc<Mutex<Vec<Player>>> = Arc::clone(&players);
+    let players_clone = Arc::clone(&players);
+    socket.on("getPlayersWithLocations", move |s: SocketRef, d: Data<Value>| {
+        get_players_with_locations(s, d, players_clone.clone())
+    });
 
-    socket.on(
-        "getPlayersWithLocations",
-        move |socket: SocketRef, Data::<Value>(data), ack: AckSender, Bin(bin)| {
-            info!("Responding with players and locations list");
-            let players: std::sync::MutexGuard<Vec<Player>> = players_clone.lock().unwrap();
-            
-            match data {
-                Value::Null => println!("Received event with null data"),
-                Value::String(s) => println!("Received event with string data: {}", s),
-                _ => println!("Received event with data: {:?}", data),
-            }
-            println!("Binary payload: {:?}", bin);
-            let players_with_locations_json = serde_json::to_value(
-                players
-                    .iter()
-                    .map(|player| json!({ 
-                        "id": player.socket.id, 
-                        "transform": player.transform.as_ref().unwrap().location
-                    }))
-                    .collect::<Vec<_>>(),
-            )
-            .unwrap();
-            info!(
-                "Players with Locations as JSON: {}",
-                players_with_locations_json
-            );
-            let players = vec![players_with_locations_json];
-            socket.emit("playersWithLocations", &players).ok();
-        },
-    );
-
-    let players_clone: Arc<Mutex<Vec<Player>>> = Arc::clone(&players);
-    socket.on("broadcastMessage", move |Data::<Value>(data), _: Bin| {
-        let players: std::sync::MutexGuard<Vec<Player>> = players_clone.lock().unwrap();
-        for player in &*players {
-            player.socket.emit("broadcastMessage", data.clone()).ok();
-        }
+    let players_clone = Arc::clone(&players);
+    socket.on("broadcastMessage", move |d: Data<Value>| {
+        broadcast_message(d, players_clone.clone())
     });
 }
